Webinspect auth (#111)
* Add new webinspect fields to config

* Create WI auth config class

* Add webinspect to admin credentials command

* Add webinspect to wipe all creds

* Fix conf test case

* Add auth to query class and commands

* Fix silly bug in WI auth config

* Remove duplicated/unused code

* Add username and password command args to WI list

* Whoops....

* Add username/password to download

* Add auth to WI proxy commands

* Attempt to fix testcases

* Add auth to WI scan command

* Add auth to JIT Scheduler

* Bump required webinspectapi version

* Add 401 handling to WI query client

* Add 401 handling to proxy client

* Add 401 error handling to WI scan client

* Add WI username/password saving

* Update docs

* Rev version

+++ webbreaker/webinspectclient.py
@@ -20,17 +20,20 @@ except (ImportError, AttributeError):  # Python3
 
 
 class WebinspectClient(object):
-    def __init__(self, webinspect_setting):
+    def __init__(self, webinspect_setting, username=None, password=None):
         Logger.app.debug("Starting webinespect client initialization")
 
         config = WebInspectConfig()
         lb = WebInspectJitScheduler(endpoints=config.endpoints,
                                     size_list=config.sizing,
-                                    size_needed=webinspect_setting['webinspect_scan_size'])
+                                    size_needed=webinspect_setting['webinspect_scan_size'], username=username, password=password)
         Logger.app.info("Querying WebInspect scan engines for availability.")
         endpoint = lb.get_endpoint()
         if not endpoint:
             raise EnvironmentError("Scheduler found no available endpoints.")
+
+        self.username = username
+        self.password = password
         self.url = endpoint
         self.settings = webinspect_setting['webinspect_settings']
         self.scan_name = webinspect_setting['webinspect_scan_name']
@@ -66,9 +69,11 @@ class WebinspectClient(object):
 
     def __settings_exists__(self):
         try:
-            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
             response = api.list_settings()
-
+            if response.response_code == 401:
+                Logger.app.critical("An Authorization Error occured.")
+                exit(1)
             if response.success:
                 for setting in response.data:
                     if setting in self.settings:
@@ -92,9 +97,11 @@ class WebinspectClient(object):
                                                                          self.start_urls, self.workflow_macros,
                                                                          self.allowed_hosts))
 
-        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
         response = api.create_scan(overrides)
-
+        if response.response_code == 401:
+            Logger.app.critical("An Authorization Error occured.")
+            exit(1)
         logger_response = json.dumps(response, default=lambda o: o.__dict__, sort_keys=True)
         Logger.app.debug("Request sent to {0}:\n{1}".format(self.url, overrides))
         Logger.app.debug("Response from {0}:\n{1}\n".format(self.url, logger_response))
@@ -118,9 +125,11 @@ class WebinspectClient(object):
         # Export scan as a xml for Threadfix or other Vuln Management System
         Logger.app.info('Exporting scan: {} as {}'.format(scan_id, extension))
         detail_type = 'Full' if extension == 'xml' else None
-        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
         response = api.export_scan_format(scan_id, extension, detail_type)
-
+        if response.response_code == 401:
+            Logger.app.critical("An Authorization Error occured.")
+            exit(1)
         if response.success:
             try:
                 with open('{0}.{1}'.format(self.scan_name, extension), 'wb') as f:
@@ -133,16 +142,22 @@ class WebinspectClient(object):
             Logger.app.error('Unable to retrieve scan results. {} '.format(response.message))
 
     def get_policy_by_guid(self, policy_guid):
-        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
         response = api.get_policy_by_guid(policy_guid)
+        if response.response_code == 401:
+            Logger.app.critical("An Authorization Error occured.")
+            exit(1)
         if response.success:
             return response.data
         else:
             return None
 
     def get_policy_by_name(self, policy_name):
-        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
         response = api.get_policy_by_name(policy_name)
+        if response.response_code == 401:
+            Logger.app.critical("An Authorization Error occured.")
+            exit(1)
         if response.success:
             return response.data
         else:
@@ -152,16 +167,22 @@ class WebinspectClient(object):
         try:
 
             if scan_name:
-                api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+                api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
                 response = api.get_scan_by_name(scan_name)
+                if response.response_code == 401:
+                    Logger.app.critical("An Authorization Error occured.")
+                    exit(1)
                 if response.success:
                     scan_guid = response.data[0]['ID']
                 else:
                     Logger.app.error(response.message)
                     return None
 
-            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
             response = api.get_scan_issues(scan_guid)
+            if response.response_code == 401:
+                Logger.app.critical("An Authorization Error occured.")
+                exit(1)
             if response.success:
                 return response.data_json(pretty=True)
             else:
@@ -173,16 +194,22 @@ class WebinspectClient(object):
         try:
 
             if scan_name:
-                api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+                api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
                 response = api.get_scan_by_name(scan_name)
+                if response.response_code == 401:
+                    Logger.app.critical("An Authorization Error occured.")
+                    exit(1)
                 if response.success:
                     scan_guid = response.data[0]['ID']
                 else:
                     Logger.app.error(response.message)
                     return None
 
-            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
             response = api.get_scan_log(scan_guid)
+            if response.response_code == 401:
+                Logger.app.critical("An Authorization Error occured.")
+                exit(1)
             if response.success:
                 return response.data_json()
             else:
@@ -191,9 +218,12 @@ class WebinspectClient(object):
             Logger.app.error("get_scan_log failed: {}".format(e))
 
     def get_scan_status(self, scan_guid):
-        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
         try:
             response = api.get_current_status(scan_guid)
+            if response.response_code == 401:
+                Logger.app.critical("An Authorization Error occured.")
+                exit(1)
             status = json.loads(response.data_json())['ScanStatus']
             return status
         except (ValueError, UnboundLocalError, TypeError) as e:
@@ -202,9 +232,11 @@ class WebinspectClient(object):
 
     def list_policies(self):
         try:
-            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
             response = api.list_policies()
-
+            if response.response_code == 401:
+                Logger.app.critical("An Authorization Error occured.")
+                exit(1)
             if response.success:
                 for policy in response.data:
                     Logger.console.info("{}".format(policy))
@@ -217,9 +249,11 @@ class WebinspectClient(object):
     def list_scans(self):
 
         try:
-            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
             response = api.list_scans()
-
+            if response.response_code == 401:
+                Logger.app.critical("An Authorization Error occured.")
+                exit(1)
             if response.success:
                 for scan in response.data:
                     Logger.console.info("{}".format(scan))
@@ -231,9 +265,11 @@ class WebinspectClient(object):
 
     def list_webmacros(self):
         try:
-            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
             response = api.list_webmacros()
-
+            if response.response_code == 401:
+                Logger.app.critical("An Authorization Error occured.")
+                exit(1)
             if response.success:
                 for webmacro in response.data:
                     Logger.console.info("{}".format(webmacro))
@@ -245,25 +281,34 @@ class WebinspectClient(object):
 
     def policy_exists(self, policy_guid):
         # true if policy exists
-        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
         response = api.get_policy_by_guid(policy_guid)
+        if response.response_code == 401:
+            Logger.app.critical("An Authorization Error occured.")
+            exit(1)
         return response.success
 
     def stop_scan(self, scan_guid):
-        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
         response = api.stop_scan(scan_guid)
+        if response.response_code == 401:
+            Logger.app.critical("An Authorization Error occured.")
+            exit(1)
         return response.success
 
     def upload_policy(self):
         # if a policy of the same name already exists, delete it prior to upload
         try:
-            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
             # bit of ugliness here. I'd like to just have the policy name at this point but I don't
             # so find it in the full path
             # TODO: Verify split here
             response = api.get_policy_by_name(ntpath.basename(self.webinspect_upload_policy).split('.')[0])
+            if response.response_code == 401:
+                Logger.app.critical("An Authorization Error occured.")
+                exit(1)
             if response.success and response.response_code == 200:  # the policy exists on the server already
-                api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+                api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
                 response = api.delete_policy(response.data['uniqueId'])
                 if response.success:
                     Logger.app.debug("Deleted policy {} from server".format(
@@ -272,9 +317,11 @@ class WebinspectClient(object):
             Logger.app.error("Verify if deletion of existing policy failed: {}".format(e))
 
         try:
-            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
             response = api.upload_policy(self.webinspect_upload_policy)
-
+            if response.response_code == 401:
+                Logger.app.critical("An Authorization Error occured.")
+                exit(1)
             if response.success:
                 Logger.console.debug("Uploaded policy {} to server.".format(self.webinspect_upload_policy))
             else:
@@ -287,9 +334,11 @@ class WebinspectClient(object):
     def upload_settings(self):
 
         try:
-            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+            api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
             response = api.upload_settings(self.webinspect_upload_settings)
-
+            if response.response_code == 401:
+                Logger.app.critical("An Authorization Error occured.")
+                exit(1)
             if response.success:
                 Logger.console.debug("Uploaded settings {} to server.".format(self.webinspect_upload_settings))
             else:
@@ -303,8 +352,11 @@ class WebinspectClient(object):
     def upload_webmacros(self):
         try:
             for webmacro in self.webinspect_upload_webmacros:
-                api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+                api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
                 response = api.upload_webmacro(webmacro)
+                if response.response_code == 401:
+                    Logger.app.critical("An Authorization Error occured.")
+                    exit(1)
                 if response.success:
                     Logger.console.debug("Uploaded webmacro {} to server.".format(webmacro))
                 else:
@@ -320,9 +372,11 @@ class WebinspectClient(object):
         :return:
         """
         # WebInspect Scan has started, wait here until it's done
-        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False)
+        api = webinspectapi.WebInspectApi(self.url, verify_ssl=False, username=self.username, password=self.password)
         response = api.wait_for_status_change(scan_id)  # this line is the blocker
-
+        if response.response_code == 401:
+            Logger.app.critical("An Authorization Error occured.")
+            exit(1)
         if response.success:
             Logger.console.debug('Scan status {}'.format(response.data))
         else:
